+++ src/renderer/parser.test.ts
@@ -10,12 +10,13 @@ describe('Unit: renderer/parser', () => {
 
   describe('escape', () => {
     it('ab', () => {
-      assert(escape('*a\\\nb*') === '\\*a\\\\\nb\\*');
+      assert(escape('*a\\\nb*') === '\\*a\\\nb\\*');
     });
 
     it('autolink', () => {
       assert(parse(escape('https://a')).firstElementChild!.innerHTML === '<a href="https://a" rel="noopener" target="_blank">https://a</a>');
-      assert(parse(escape('!https://host')).firstElementChild!.innerHTML === '<a href="https://host" rel="noopener" target="_blank"><img data-src="https://host" alt=""></a>');
+      assert(parse(escape('https://[::]')).firstElementChild!.innerHTML === '<a href="https://[::]" rel="noopener" target="_blank">https://[::]</a>');
+      assert(parse(escape('!https://a')).firstElementChild!.innerHTML === '<a href="https://a" rel="noopener" target="_blank"><img data-src="https://a" alt=""></a>');
       assert(parse(escape('@a')).firstElementChild!.innerHTML === '<span class="account">@a</span>');
     });